Extract separate global deprecation helpers into standalone module.
No logical changes here, just moving the code to detect and generate the
message out into a stand alone module (makes it easier to reason about
and possible to unit test).
